+++ relay-profiling/src/sample.rs
@@ -175,9 +175,11 @@ impl SampleProfile {
                     && self.device.manufacturer.is_some()
                     && self.device.model.is_some()
             }
-            Platform::Python | Platform::Javascript | Platform::Node | Platform::Php => {
-                self.runtime.is_some()
-            }
+            Platform::Dotnet
+            | Platform::Javascript
+            | Platform::Node
+            | Platform::Php
+            | Platform::Python => self.runtime.is_some(),
             _ => true,
         }
     }