+++ CHANGELOG.md
@@ -7,8 +7,7 @@
 - Extract attachments from transaction events and send them to kafka individually. ([#1844](https://github.com/getsentry/relay/pull/1844))
 - Protocol validation for source map image type. ([#1869](https://github.com/getsentry/relay/pull/1869))
 - Strip quotes from client hint values. ([#1874](https://github.com/getsentry/relay/pull/1874))
-- Add PHP support for profiling. ([#1871](https://github.com/getsentry/relay/pull/1871))
-- Add Javascript support for profiling. ([#1876](https://github.com/getsentry/relay/pull/1876))
+- Add Dotnet, Javascript and PHP support for profiling. ([#1871](https://github.com/getsentry/relay/pull/1871), [#1876](https://github.com/getsentry/relay/pull/1876), [#1885](https://github.com/getsentry/relay/pull/1885))
 
 **Bug Fixes**: